convert to scanline printing
apparently writing formatted characters to the terminal stream is really
really slow. this is a first attempt to write fewer characters per frame

this seems to stabilize the frame rate, but is still a long ways from
feeling responsive.

next step may be to go to a full matrix printer, and look to emit only
characters on a scanline that have changed in the last frame